+++ src/__tests__/array.test.ts
@@ -1,4 +1,4 @@
-import { JsgArray } from './array'
+import { JsgArray } from '../array'
 
 describe('JsgArray', () => {
   describe('constructor', () => {
